chore(editor) renamed some test files to please dependency-cruiser
